Trocado método user por findUser
-- Substituído método user da userApi pelo método 'findUser', que tem 3 métodos internos 'byId', 'byEmail' e 'byCookie', que procuram o usuário e retornam a classe User com ele se encontrarem
-- Separado clase User em um arquivo separado
-- Adaptado chamadas para user findUser.byEmail onde antes era utilizado user
