+++ src/userApi/user.ts
@@ -1,0 +1,59 @@
+import { sha512 } from 'js-sha512'
+import { Person } from '../db/models/person/interface'
+
+export default class User {
+	/**
+	 * Cria o sha512 do salt com o password, seguindo o padrão do createUser
+	 */
+	private _hashPassword = (password: string): string =>
+		sha512.create()
+			.update(this.person.credentials.salt)
+			.update(password)
+			.hex()
+
+	constructor(person: Person) {
+		this.person = person
+	}
+
+	/**
+	 * Documento do mongodb desta person, acessível publicamente
+	 */
+	person: Person
+
+	/**
+	 * Retorna o saldo de um usuário para determinada currency
+	 */
+	getBalance = (currency: string): number => this.person.currencies[currency].balance
+
+	/**
+	 * Retorna as accounts de um usuário para determinada currency
+	 */
+	getAccounts = (currency: string): string[] => this.person.currencies[currency].accounts
+
+	/**
+	 * Retorna o objectId do documento person
+	 */
+	getObjectId = (): Person['_id'] => this.person._id
+
+	/**
+	 * Retorna 'void' se o password informado é o password correto do usuário
+	 * 
+	 * @throws InvalidPassword if password is invalid
+	 */
+	checkPassword = (password: string): void => {
+		const password_hash = this._hashPassword(password)
+		if (password_hash != this.person.credentials.password_hash)
+			throw 'InvalidPassword'
+	}
+
+	/**
+	 * Atualiza o password do usuário com o password informado
+	 * 
+	 * @returns The updated person object
+	 */
+	changePassword = async (password: string): Promise<Person> => {
+		this.person.credentials.password_hash = this._hashPassword(password)
+		await this.person.save()
+		return this.person
+	}
+}
